Conflicts:
	webapps/administrator/app/templates/user.hbs

+++ rule/Task/Resources/scripts/task/rest/TagService.java
@@ -15,51 +15,57 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
-import task.dao.TaskDAO;
-import task.model.Task;
-import task.validation.TaskValidator;
-import task.validation.ValidationError;
-
 import com.fasterxml.jackson.annotation.JsonRootName;
 import com.flabser.restful.RestProvider;
 
+import task.dao.TagDAO;
+import task.model.Tag;
+import task.pojo.Errors;
+import task.validation.TagValidator;
+import task.validation.ValidationError;
 
-@Path("tasks")
-public class TaskService extends RestProvider {
 
-	private TaskValidator validator = new TaskValidator();
+@Path("tags")
+public class TagService extends RestProvider {
+
+	private TagValidator validator = new TagValidator();
 
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response get() {
-		TaskDAO dao = new TaskDAO(getSession());
-		return Response.ok(new Tasks(dao.findAll())).build();
+		TagDAO dao = new TagDAO(getSession());
+		return Response.ok(new Tags(dao.findAll())).build();
 	}
 
 	@GET
-	@Path("{id}")
+	@Path("/{id}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response get(@PathParam("id") long id) {
-		TaskDAO dao = new TaskDAO(getSession());
-		Task m = dao.findById(id);
-		//
+		TagDAO dao = new TagDAO(getSession());
+		Tag m = dao.findById(id);
 		if (m == null) {
 			return Response.noContent().status(Status.NOT_FOUND).build();
 		}
-		//
+
 		return Response.ok(m).build();
 	}
 
 	@POST
 	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response create(Task m) {
+	public Response create(Tag m) {
+		TagDAO dao = new TagDAO(getSession());
+
+		// parent tag
+		if (m.getParent() != null) {
+			m.setParent(dao.findById(m.getParent().getId()));
+		}
+
 		ValidationError ve = validator.validate(m);
 		if (ve.hasError()) {
 			return Response.status(Status.BAD_REQUEST).entity(ve).build();
 		}
 
-		TaskDAO dao = new TaskDAO(getSession());
 		return Response.ok(dao.add(m)).build();
 	}
 
@@ -67,23 +73,29 @@ public class TaskService extends RestProvider {
 	@Path("/{id}")
 	@Produces(MediaType.APPLICATION_JSON)
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response update(@PathParam("id") long id, Task m) {
+	public Response update(@PathParam("id") long id, Tag m) {
+		TagDAO dao = new TagDAO(getSession());
+
 		m.setId(id);
+
+		// parent tag
+		if (m.getParent() != null) {
+			m.setParent(dao.findById(m.getParent().getId()));
+		}
 
 		ValidationError ve = validator.validate(m);
 		if (ve.hasError()) {
 			return Response.status(Status.BAD_REQUEST).entity(ve).build();
 		}
 
-		TaskDAO dao = new TaskDAO(getSession());
-		Task pm = dao.findById(id);
-		//
+		Tag pm = dao.findById(id);
 		if (pm == null) {
 			return Response.noContent().status(Status.NOT_FOUND).build();
 		}
-		//
+
 		pm.setName(m.getName());
-		//
+		pm.setParent(m.getParent());
+
 		return Response.ok(dao.update(pm)).build();
 	}
 
@@ -91,18 +103,27 @@ public class TaskService extends RestProvider {
 	@Path("/{id}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response delete(@PathParam("id") long id) {
-		TaskDAO dao = new TaskDAO(getSession());
-		Task m = dao.findById(id);
-		dao.delete(m);
+		TagDAO dao = new TagDAO(getSession());
+		Tag m = dao.findById(id);
+		if (m != null) {
+			if (dao.existsChildCategory(m)) {
+				Errors msg = new Errors();
+				msg.setMessage("exists_child");
+				return Response.status(Status.BAD_REQUEST).entity(msg).build();
+			} else {
+				dao.delete(m);
+			}
+		}
+
 		return Response.status(Status.NO_CONTENT).build();
 	}
 
-	@JsonRootName("tasks")
-	class Tasks extends ArrayList <Task> {
+	@JsonRootName("tags")
+	class Tags extends ArrayList <Tag> {
 
 		private static final long serialVersionUID = 1L;
 
-		public Tasks(Collection <? extends Task> m) {
+		public Tags(Collection <? extends Tag> m) {
 			addAll(m);
 		}
 	}
